Extract test helper class for reuse
