+++ testing/trino-tests/src/test/java/io/trino/connector/informationschema/TestInformationSchemaConnector.java
@@ -13,20 +13,16 @@
  */
 package io.trino.connector.informationschema;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMultiset;
 import com.google.common.collect.Multiset;
 import io.trino.Session;
-import io.trino.connector.MockConnectorFactory;
 import io.trino.plugin.tpch.TpchPlugin;
-import io.trino.spi.Plugin;
-import io.trino.spi.TrinoException;
-import io.trino.spi.connector.ConnectorFactory;
 import io.trino.testing.AbstractTestQueryFramework;
 import io.trino.testing.CountingMockConnector;
 import io.trino.testing.DistributedQueryRunner;
 import io.trino.testng.services.ManageTestResources;
+import io.trino.tests.FailingMockConnectorPlugin;
 import org.intellij.lang.annotations.Language;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
@@ -36,7 +32,6 @@ import org.junit.jupiter.api.Timeout;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-import static io.trino.spi.StandardErrorCode.GENERIC_INTERNAL_ERROR;
 import static io.trino.testing.MultisetAssertions.assertMultisetsEqual;
 import static io.trino.testing.TestingSession.testSessionBuilder;
 import static java.util.stream.Collectors.joining;
@@ -427,36 +422,5 @@ public class TestInformationSchemaConnector
         });
 
         assertMultisetsEqual(actualMetadataCallsCount, expectedMetadataCallsCount);
-    }
-
-    private static final class FailingMockConnectorPlugin
-            implements Plugin
-    {
-        @Override
-        public Iterable<ConnectorFactory> getConnectorFactories()
-        {
-            return ImmutableList.of(
-                    MockConnectorFactory.builder()
-                            .withName("failing_mock")
-                            .withListSchemaNames(session -> {
-                                throw new TrinoException(GENERIC_INTERNAL_ERROR, "Catalog is broken");
-                            })
-                            .withListTables((session, schema) -> {
-                                throw new TrinoException(GENERIC_INTERNAL_ERROR, "Catalog is broken");
-                            })
-                            .withGetViews((session, prefix) -> {
-                                throw new TrinoException(GENERIC_INTERNAL_ERROR, "Catalog is broken");
-                            })
-                            .withGetMaterializedViews((session, prefix) -> {
-                                throw new TrinoException(GENERIC_INTERNAL_ERROR, "Catalog is broken");
-                            })
-                            .withListTablePrivileges((session, prefix) -> {
-                                throw new TrinoException(GENERIC_INTERNAL_ERROR, "Catalog is broken");
-                            })
-                            .withStreamTableColumns((session, prefix) -> {
-                                throw new TrinoException(GENERIC_INTERNAL_ERROR, "Catalog is broken");
-                            })
-                            .build());
-        }
     }
 }
